Fixed prototypes almost everywhere.
Rev: src/acconfig.h:1.9
Rev: src/array.c:1.12
Rev: src/array.h:1.4
Rev: src/backend.c:1.15
Rev: src/backend.h:1.3
Rev: src/builtin_functions.c:1.41
Rev: src/builtin_functions.h:1.4
Rev: src/callback.c:1.9
Rev: src/callback.h:1.4
Rev: src/constants.c:1.12
Rev: src/constants.h:1.5
Rev: src/docode.c:1.20
Rev: src/docode.h:1.3
Rev: src/dynamic_load.c:1.19
Rev: src/dynamic_load.h:1.2
Rev: src/error.c:1.8
Rev: src/error.h:1.6
Rev: src/gc.c:1.12
Rev: src/gc.h:1.10
Rev: src/hashtable.c:1.3
Rev: src/hashtable.h:1.4
Rev: src/interpret.c:1.45
Rev: src/interpret.h:1.12
Rev: src/language.yacc:1.46
Rev: src/las.c:1.36
Rev: src/las.h:1.8
Rev: src/lex.c:1.26
Rev: src/lex.h:1.5
Rev: src/main.c:1.23
Rev: src/mapping.c:1.21
Rev: src/mapping.h:1.4
Rev: src/modules/Gdbm/gdbmmod.c:1.4
Rev: src/modules/Gmp/mpz_glue.c:1.20
Rev: src/modules/Image/image.c:1.36
Rev: src/modules/Image/polyfill.c:1.5
Rev: src/modules/Mysql/mysql.c:1.6
Rev: src/modules/Pipe/pipe.c:1.11
Rev: src/modules/Regexp/glue.c:1.7
Rev: src/modules/Regexp/pike_regexp.c:1.5
Rev: src/modules/Ssleay/ssleay.c:1.7
Rev: src/modules/Yp/yp.c:1.8
Rev: src/modules/_Crypto/lib/desQuick.c:1.3
Rev: src/modules/call_out/call_out.c:1.13
Rev: src/modules/files/efuns.c:1.33
Rev: src/modules/files/file.c:1.51
Rev: src/modules/files/file.h:1.5
Rev: src/modules/files/socket.c:1.14
Rev: src/modules/math/math.c:1.5
Rev: src/modules/spider/accesseddb.c:1.10
Rev: src/modules/spider/dumudp.c:1.14
Rev: src/modules/spider/spider.c:1.39
Rev: src/modules/spider/streamed_parser.c:1.7
Rev: src/modules/spider/streamed_parser.h:1.2
Rev: src/multiset.c:1.7
Rev: src/multiset.h:1.3
Rev: src/object.c:1.21
Rev: src/object.h:1.11
Rev: src/opcodes.c:1.8
Rev: src/opcodes.h:1.3
Rev: src/operators.c:1.16
Rev: src/operators.h:1.3
Rev: src/peep.c:1.15
Rev: src/peep.h:1.3
Rev: src/pike_types.c:1.25
Rev: src/pike_types.h:1.5
Rev: src/program.c:1.37
Rev: src/program.h:1.18
Rev: src/rusage.c:1.8
Rev: src/rusage.h:1.2
Rev: src/signal_handler.c:1.14
Rev: src/signal_handler.h:1.3
Rev: src/stralloc.c:1.17
Rev: src/stralloc.h:1.9
Rev: src/svalue.c:1.14
Rev: src/threads.c:1.24
Rev: src/threads.h:1.12

+++ src/modules/Yp/yp.c
@@ -143,7 +143,7 @@ void f_map(INT32 args)
 static void f_order(INT32 args)
 {
   int err;
-  unsigned long ret;
+  unsigned int ret;
   check_all_args("yp->order", args, BIT_STRING, 0);
   
   err = yp_order( this->domain, sp[-args].u.string->str, &ret);
